mcp-weather-ink-suite/server-pc/pyproject.toml
[project]
name = "mcp-weather-air-info-server"
version = "0.1.0"
description = "MCP Server for Weather and Air Quality Info"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "httpx",
    "pydantic",
    "python-dotenv",
    "pytz",
]
